refactor(gui): updated methods for auto registration groups
* The methods for auto registration groups were moved to the registrarManager, so the methods urls
had to be updated.
