+++ perun-web-gui/src/main/java/cz/metacentrum/perun/webgui/tabs/registrartabs/AddAutoRegGroupTabItem.java
@@ -19,9 +19,9 @@ import cz.metacentrum.perun.webgui.client.resources.PerunSearchEvent;
 import cz.metacentrum.perun.webgui.client.resources.SmallIcons;
 import cz.metacentrum.perun.webgui.json.JsonCallbackEvents;
 import cz.metacentrum.perun.webgui.json.JsonUtils;
-import cz.metacentrum.perun.webgui.json.groupsManager.AddAutoRegGroups;
+import cz.metacentrum.perun.webgui.json.registrarManager.AddAutoRegGroups;
 import cz.metacentrum.perun.webgui.json.groupsManager.GetAllGroups;
-import cz.metacentrum.perun.webgui.json.groupsManager.GetAutoRegistrationGroups;
+import cz.metacentrum.perun.webgui.json.registrarManager.GetAutoRegistrationGroups;
 import cz.metacentrum.perun.webgui.model.Group;
 import cz.metacentrum.perun.webgui.model.VirtualOrganization;
 import cz.metacentrum.perun.webgui.tabs.TabItem;
