+++ perun-web-gui/src/main/java/cz/metacentrum/perun/webgui/tabs/registrartabs/AutoRegistrationGroupsTabItem.java
@@ -18,9 +18,8 @@ import cz.metacentrum.perun.webgui.client.resources.PerunSearchEvent;
 import cz.metacentrum.perun.webgui.client.resources.SmallIcons;
 import cz.metacentrum.perun.webgui.json.GetEntityById;
 import cz.metacentrum.perun.webgui.json.JsonCallbackEvents;
-import cz.metacentrum.perun.webgui.json.groupsManager.DeleteGroups;
-import cz.metacentrum.perun.webgui.json.groupsManager.GetAutoRegistrationGroups;
-import cz.metacentrum.perun.webgui.json.groupsManager.RemoveGroupsFromAutoRegistration;
+import cz.metacentrum.perun.webgui.json.registrarManager.GetAutoRegistrationGroups;
+import cz.metacentrum.perun.webgui.json.registrarManager.RemoveGroupsFromAutoRegistration;
 import cz.metacentrum.perun.webgui.model.Group;
 import cz.metacentrum.perun.webgui.model.VirtualOrganization;
 import cz.metacentrum.perun.webgui.tabs.RegistrarTabs;
